[ECLIPSE CLEANUP] Imports, brackets, annotations, casts, serialVersionUIDs!
git-svn-id: file:///home/svn/omero/trunk@1165 05709c45-44f0-0310-885b-81a1db45b4a6

+++ components/common/test/ome/system/utests/VersionTest.java
@@ -18,7 +18,7 @@ public class VersionTest extends TestCase {
     }
 
     public void testGetRevision() throws Exception {
-        assertEquals((Integer) 100, (Integer) Version.getRevision(Blank.class));
+        assertEquals((Integer) 100, Version.getRevision(Blank.class));
     }
 
     public void testGetDate() throws Exception {
